+++ tests/Jroc.Tests/Math/JavaScript/Math_DefinePropertyReplacement_Dispatch.js
@@ -1,0 +1,9 @@
+Object.defineProperty(Math, "sqrt", {
+  value: function () {
+    return 789;
+  },
+  configurable: true,
+  writable: true
+});
+
+console.log(Math.sqrt(9));